Zmiana funkcji double na float (#1)

+++ src/org/lauchcode/Main.java
@@ -15,17 +15,17 @@ public class Main {
         System.out.println("Podaj Swój wiek:");
         String inputAge;
         inputAge = scanner.next();
-        double procesAge = Double.parseDouble(inputAge);
-        String outAge = String.format ("%.1f", procesAge);
+        float procesAge = Float.parseFloat(inputAge);
+        String outAge = String.format ("%.0f", procesAge);
         System.out.println("Podaj Swóją wzrost:");
         String inputTall;
         inputTall = scanner.next();
-        double procesTall = Double.parseDouble(inputTall); // Test github
+        float procesTall = Float.parseFloat(inputTall); // Test github
         String outTall = String.format ("%.1f", procesTall);
         System.out.println("Podaj Swóją wagę:");
         String inputWeight;
         inputWeight = scanner.next();
-       double procesWeight = Double.parseDouble(inputWeight);
+        float procesWeight = Float.parseFloat(inputWeight);
         String outWeight = String.format ("%.1f", procesWeight);
         System.out.println("Nazywasz się : "+inputName+" "+inputSurname+",masz "+outAge+" lat, masz "+outTall+ " cm wzrosu oraz ważysz "+ outWeight+ " kg" );
